% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Instantiates a set of cohorts form a folder containing the JSON cohort definitions}
\usage{
createCohorts(
  jsonfileLocation,
  cohortsToCreate,
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  saveDirectory
)
}
\arguments{
\item{jsonfileLocation, }{}

\item{connectionDetails, }{}

\item{cdmDatabaseSchema, }{}

\item{cohortDatabaseSchema, }{}

\item{cohortTable}{}
}
\description{
Instantiates a set of cohorts form a folder containing the JSON cohort definitions
}
\details{
The user specifies the location of the JSON definitions as well as the connection details for the data abse
}
