% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{createSampleTrainSetSettings}
\alias{createSampleTrainSetSettings}
\title{Create sample settings to sample the train set.}
\usage{
createSampleTrainSetSettings(
  numberTrainSetOutcomes = NULL,
  sampleSeed = sample(10000, 1)
)
}
\arguments{
\item{numberTrainSetOutcomes}{The absolute number of desired number of outcomes to retain in the train set.}

\item{sampleSeed}{A seed to use when sampling.}
}
\description{
Given a train set and a number of required outcomes, the train set outcomes are reduced to the number of required train outcomes, trying to maintain the original outcome to non-outcome distribution.
}
\details{
Create a sampleSettings object containing a number of required outcomes to reduce the train set sample, maintaining the original outcome to non-outcome distribution.
}
