% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eunomiaDesigns}
\alias{eunomiaDesigns}
\title{eunomiaDesigns}
\format{
An object of class \code{list} of length 5.
}
\usage{
eunomiaDesigns
}
\description{
Patient-Level Prediction Model Designs for the Eunomia dataset
}
\details{
A prespecified list containing model designs with class `modelDesign` as generetated by `PatientLevelPrediction::createModelDesign`
See \link[PatientLevelPrediction]{createModelDesign} for more details and \url{https://ohdsi.github.io/PatientLevelPrediction/articles/BuildingMultiplePredictiveModels.html#creating-a-model-design} on examples how to create a model design.
}
\keyword{datasets}
