% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eunomiaTasks}
\alias{eunomiaTasks}
\title{eunomiaTasks}
\format{
A data frame with 7 variables:
\describe{
\item{\code{pId}}{Problem Id}
\item{\code{analysisName}}{Name given for this analysis}
\item{\code{problemSpecification}}{Problem specification}
\item{\code{TAR_start_day}}{Starting day for time-at-risk}
\item{\code{TAR_end_day}}{Ending day for time-at-risk}
\item{\code{targetId}}{The id of the target cohort as found in `inst/settings/cohortsToCreate.csv`}
\item{\code{outcomeId}}{The id of the outcome cohort as found in `inst/settings/cohortsToCreate.csv`}
}

To create the cohorts for the Eunomia dataset we can call \link[Eunomia]{createCohorts}
}
\usage{
eunomiaTasks
}
\description{
Task Specification for the Eunomia dataset
}
\keyword{datasets}
