% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBenchmarkData.R
\name{extractBenchmarkData}
\alias{extractBenchmarkData}
\title{Extract the data for the benchmark}
\usage{
extractBenchmarkData(benchmarkDesign, createStudyPopulation = TRUE)
}
\arguments{
\item{benchmarkDesign}{An object of class \code{benchmarkDesign}.}

\item{createStudyPopulation}{Whether to create the study population or not.}
}
\description{
Extracts the data needed to run the benchmark models and optionally creates the study population.
}
