% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBenchmarkModelPerformance.R
\name{getBenchmarkModelPerformance}
\alias{getBenchmarkModelPerformance}
\title{Get the mode performance.}
\usage{
getBenchmarkModelPerformance(benchmarkDesign)
}
\arguments{
\item{benchmarkDesign}{An object of class \code{benchmarkDesign}.}
}
\description{
Collect the performance of each model specified in the benchmark design.
}
