% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBenchmarkTasks.R
\name{loadModelDesigns}
\alias{loadModelDesigns}
\title{Loads the model design list}
\usage{
loadModelDesigns(designs = "all")
}
\description{
Loads the models design list.
}
\details{
A model design is an object of class `modelDesign` as generated by \code{PatientLevelPrediction::createModelDesign()}.


#' @param designs Define which designs should be loaded. Choice of "all" to load all tasks or a numeric vector to define which.
}
