% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelDesigns}
\alias{modelDesigns}
\title{modelDesigns}
\format{
An object of class \code{list} of length 13.
}
\usage{
modelDesigns
}
\description{
Patient-Level Prediction Model Designs for Real-World Data
}
\details{
A prespecified list containing model designs with class `modelDesign` as generetated by `PatientLevelPrediction::createModelDesign`
See \link[PatientLevelPrediction]{createModelDesign} for more details and \url{https://ohdsi.github.io/PatientLevelPrediction/articles/BuildingMultiplePredictiveModels.html#creating-a-model-design} on examples how to create a model design.
}
\keyword{datasets}
