% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLPBenchmarks-package.R
\docType{package}
\name{PLPBenchmarks-package}
\alias{PLPBenchmarks}
\alias{PLPBenchmarks-package}
\title{PLPBenchmarks: Benchmarks of clinical prediction models}
\description{
A package to execute benchmarking of clinical prediction models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mi-erasmusmc.github.io/PlpBenchmarks/}
}

}
\keyword{internal}
