% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampling.R
\name{reduceTrainSetOutcomes}
\alias{reduceTrainSetOutcomes}
\title{Sample the train set.}
\usage{
reduceTrainSetOutcomes(trainData, sampleSettings)
}
\arguments{
\item{trainData}{An object of class `trainData`.}

\item{sampleSettings}{An object of class `sampleSettings`.}
}
\description{
Given a train set and a number of required outcomes, the train set outcomes are reduced to the number of required train outcomes, trying to maintain the original outcome to non-outcome distribution.
}
\details{
Provide a number of required outcomes, to reduce the number the train set sample, maintaining the original outcome to non-outcome distribution.
}
