% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarkDesign.R
\name{runBenchmarkDesign}
\alias{runBenchmarkDesign}
\title{Run the benchmark models.}
\usage{
runBenchmarkDesign(benchmarkDesign)
}
\arguments{
\item{benchmarkDesign}{An object of class \code{benchmarkDesign}.}
}
\description{
Run the models specified in the benchmark design.
}
