% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBenchmarkModel.R
\name{runBenchmarkModel}
\alias{runBenchmarkModel}
\title{Trains one of the benchmark prediction problem with standard settingsfor a Lasso LR model}
\usage{
runBenchmarkModel(
  databaseDetails = createDatabaseDetails(),
  modelDesignList = list(createModelDesign(targetId = 1, outcomeId = 2, modelSettings =
    setLassoLogisticRegression()), createModelDesign(targetId = 1, outcomeId = 3,
    modelSettings = setLassoLogisticRegression())),
  saveDirectory = getwd(),
  sqliteLocation = file.path(saveDirectory, "sqlite")
)
}
\arguments{
\item{targetId}{The targetId of the cohort, this is available in the cohort_table of the database where the cohorts were instantiated}

\item{outcomeId}{The outcomeId of the cohort, this is available in the cohort_table of the database where the cohorts were instantiate}

\item{analysisId}{The analysisId which matches the benchmark problem}

\item{analysisName}{the analysisName which matches the benchmark problem}

\item{riskWindowStart}{The start of the time at risk period}

\item{riskWindowEnd}{The end of the time at risk period}

\item{connectionDetails}{databaseconnector connectionDetails object}

\item{cdmDatabaseSchema}{the name of the table where the cdm is found in the dtaabase}

\item{cohortDatabaseSchema}{a writable table where the cohorts are found}

\item{cohortTable}{the name of the table where the cohorts are stored}

\item{saveLoc}{location to save the model too}

\item{populationSettins}{An object created using PatientLevelPrediction::createPopulationsettings()}
}
\value{
The trained benchmark model
}
\description{
Trains one of the benchmark prediction problem with standard settingsfor a Lasso LR model
}
\details{
The user specifies the target and outcome pair as well as the TAR. Combining this with connection details and pre instantiated cohorts allows for the running of the benchmark model
}
