% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasks}
\alias{tasks}
\title{tasks}
\format{
A data frame with 13 variables:
\describe{
\item{\code{pId}}{Problem Id}
\item{\code{analysisName}}{Name given for this analysis}
\item{\code{problemSpecification}}{Problem specification}
\item{\code{TAR_start_day}}{Starting day for time-at-risk}
\item{\code{TAR_end_day}}{Ending day for time-at-risk}
\item{\code{targetId}}{The id of the target cohort as found in `inst/settings/cohortsToCreate.csv`}
\item{\code{outcomeId}}{The id of the outcome cohort as found in `inst/settings/cohortsToCreate.csv`}
\item{\code{targetCohortName}}{The name of the target cohort as found in `inst/settings/cohortsToCreate.csv`}
\item{\code{outcomeCohortName}}{The name of the outcome cohort as found in `inst/settings/cohortsToCreate.csv`}
\item{\code{targetCohortJson}}{The name of the json file containing the target cohort definition}
\item{\code{outcomeCohortJson}}{The name of the json file containing the outcome cohort definition}
}
}
\usage{
tasks
}
\description{
Task Specification for Real-World Data
}
\keyword{datasets}
