% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBenchmarkPlpResults.R
\name{viewBenchmarkPlpResults}
\alias{viewBenchmarkPlpResults}
\title{View a shiny app with the benchmark results}
\usage{
viewBenchmarkPlpResults(
  resultLocation,
  benchmarkDesign,
  cohortDefinitions = NULL,
  databaseList,
  createPlpResultTables = TRUE,
  viewShiny = TRUE,
  sqliteLocation = NULL
)
}
\arguments{
\item{resultLocation}{File path to the folder storing the results to upload.}

\item{benchmarkDesign}{A list of class `benchmarkDesign` as generated by \code{createBenchmarkDesign()}.}

\item{cohortDefinitions}{An object of class `cohortDefinitionSet` as provided by \code{CohortGenerator::getCohortDefinitionSet()}}

\item{databaseList}{A list containing the database names.}

\item{createPlpResultTables}{A logical to indicate whether to create the PlpResults tables. Defaults to `TRUE` and this will delete the tables (if they exist) and re-create them.}

\item{viewShiny}{Whether to view the shiny app or not, defaults to `TRUE`.}

\item{sqliteLocation}{Directory to save cohort counts and/or which cohorts have been generated.}
}
\description{
View a shiny app with the benchmark results
}
\details{
Upload the results to a database, create a shiny app and view it.
}
