% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBenchmarkSettings.R
\name{viewBenchmarkSettings}
\alias{viewBenchmarkSettings}
\title{View settings of a benchmark design}
\usage{
viewBenchmarkSettings(benchmarkDesign)
}
\arguments{
\item{benchmarkDesign}{A list of class \code{modelDesign} created using 
\code{PLPBenchmarks::createBenchmarkDesign()}}
}
\description{
Provides the settings as a dataframe for side-to-side comparissons.
}
