% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBenchmarkCohorts.R
\name{createBenchmarkCohorts}
\alias{createBenchmarkCohorts}
\title{Instantiates a set of cohorts form a folder containing the JSON cohort definitions}
\usage{
createBenchmarkCohorts(
  jsonfileLocation = NULL,
  cohortsToCreate = NULL,
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  saveDirectory
)
}
\arguments{
\item{jsonfileLocation}{The location of the json files for the cohorts to be created. Either \code{jsonFileLocation} or \code{cohortdToCreate} need to be specified.}

\item{cohortsToCreate}{A data frame with a column named jsonLocation. Either \code{jsonFileLocation} or \code{cohortdToCreate} need to be specified.}

\item{connectionDetails}{The connection details object to connect to the database}

\item{cdmDatabaseSchema}{The schema name where the data is located.}

\item{cohortDatabaseSchema}{The schema where the cohort's will be stored.}

\item{cohortTable}{The name of the table under which the cohort/s will be created}

\item{saveDirectory}{Directory to save cohort counts.}
}
\description{
Instantiates a set of cohorts form a folder containing the JSON cohort definitions
}
\details{
The user specifies the location of the JSON definitions as well as the connection details for the data abse
}
