% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisDesign.R
\name{createBenchmarkDesign}
\alias{createBenchmarkDesign}
\title{Create a benchmark analysis design list}
\usage{
createBenchmarkDesign(
  modelDesign,
  databaseDetails,
  requiredTrainPositiveEvents = NULL,
  covariateComparisson = c("demographics", "demographics and conditions",
    "demographics and drugs", "demographics and procedures",
    "demographics, conditions and drugs", "demographics, conditions and procedures",
    "demographics, drugs and procedures",
    "demographics, conditions, drugs and procedures"),
  analysisNames,
  saveDirectory
)
}
\arguments{
\item{covariateComparisson}{Options: \code{"demographics", "demographics and conditions", "demographics and drugs", "demographics and procedures", 
"demographics, conditions and drugs", "demographics, conditions and procedures", "demographics, drugs and procedures", 
"demographics, conditions, drugs and procedures"}}
}
\description{
Create a benchmark analysis design list
}
